Fix vpickve2gr/xvpickve2gr

+++ code/vpickve2gr_b.h
@@ -1,1 +1,1 @@
-dst = a.byte[idx];
+dst = (s8)a.byte[idx];
