+++ code/vpickve2gr_d.h
@@ -1,1 +1,1 @@
-dst = a.dword[idx];
+dst = (s64)a.dword[idx];
