+++ code/vpickve2gr_h.h
@@ -1,1 +1,1 @@
-dst = a.half[idx];
+dst = (s16)a.half[idx];
