+++ code/xvpickve2gr_w.h
@@ -1,1 +1,1 @@
-dst = a.word[idx];
+dst = (s32)a.word[idx];
